app/requirements.txt
requests
PyYAML
tornado>=4.5.1
